+++ SlothyTerminal/Services/UsageKeychainStore.swift
@@ -25,6 +25,7 @@ enum UsageKeychainStore {
       kSecAttrAccount as String: account,
       kSecValueData as String: data,
       kSecAttrAccessible as String: kSecAttrAccessibleWhenUnlockedThisDeviceOnly,
+      kSecUseDataProtectionKeychain as String: true,
     ]
 
     let status = SecItemAdd(query as CFDictionary, nil)
@@ -65,6 +66,7 @@ enum UsageKeychainStore {
       kSecAttrAccount as String: account,
       kSecReturnData as String: true,
       kSecMatchLimit as String: kSecMatchLimitOne,
+      kSecUseDataProtectionKeychain as String: true,
     ]
 
     var result: AnyObject?
@@ -101,6 +103,7 @@ enum UsageKeychainStore {
       kSecClass as String: kSecClassGenericPassword,
       kSecAttrService as String: serviceName,
       kSecAttrAccount as String: account,
+      kSecUseDataProtectionKeychain as String: true,
     ]
 
     let status = SecItemDelete(query as CFDictionary)